refactor(register): organize types, validation, and service logic
