--- src/components/Register/types.ts
@@ -1,73 +1,0 @@
-import { FormikProps } from 'formik';
-
-type RegisterFormProps = Pick<
-  FormikProps<RegisterFormFields>,
-  'values' | 'errors' | 'touched' | 'handleChange' | 'handleBlur'
-> & { setFieldValue: (field: string, value: unknown) => void };
-
-type PersonalInfoProps = Pick<
-  FormikProps<PersonalInfoFieldsValues>,
-  'values' | 'errors' | 'touched' | 'handleChange' | 'handleBlur'
-> & {
-  setFieldValue: (field: string, value: unknown) => void;
-  withEmail?: boolean;
-};
-
-type UserAddress = {
-  country: string;
-  city: string;
-  streetName: string;
-  postalCode: string;
-  isDefault: boolean;
-  useSame?: boolean;
-  [key: string]: string | boolean | undefined;
-};
-
-type RegisterFormFields = {
-  email: string;
-  password: string;
-  confirmPassword: string;
-  firstName: string;
-  lastName: string;
-  dateOfBirth: string;
-  phoneNumber: string;
-  billingAddress: UserAddress;
-  shippingAddress: UserAddress;
-};
-
-type CommonFormProps = {
-  name: string;
-  label: string;
-  type?: string;
-  placeholder?: string;
-  value: string | boolean | undefined;
-  onChange: React.ChangeEventHandler<HTMLElement>;
-  onBlur: React.FocusEventHandler<HTMLElement>;
-  error?: string;
-  touched?: boolean;
-  withToggle?: boolean;
-  show?: boolean;
-  onToggle?: () => void;
-  withDatePicker?: boolean;
-  autoComplete?: string;
-  onDatePick?: (value: string) => void;
-  asSelect?: boolean;
-  options?: { value: string; label: string }[];
-};
-
-type PersonalInfoFieldsValues = {
-  firstName: string;
-  lastName: string;
-  dateOfBirth: string;
-  phoneNumber: string;
-  email?: string;
-};
-
-export type {
-  RegisterFormProps,
-  UserAddress,
-  RegisterFormFields,
-  CommonFormProps,
-  PersonalInfoFieldsValues,
-  PersonalInfoProps
-};
